Demographics table and lasso
Demographics table was added to the manuscript and the first round of lasso was fitted. However, for some reason the estimate for all of the predictors chosen by LASSO were equal to zero. I am not sure if this is because the penalty is too high or why else. I will try and research this.